refactor: e2e 테스트 명칭 변경
